+++ InteractionNew/Runtime/Scripts/GenericInteraction/GenericInteractable.cs
@@ -3,10 +3,10 @@ using Reflectis.SDK.Platform;
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
-
+using Unity.VisualScripting;
 using UnityEditor;
-
 using UnityEngine;
 
 namespace Reflectis.SDK.InteractionNew
@@ -42,11 +42,13 @@ namespace Reflectis.SDK.InteractionNew
         [SerializeField] private List<AwaitableScriptableAction> onInteractActions = new();
         [SerializeField] private List<AwaitableScriptableAction> onInteractFinishActions = new();
 
+        [SerializeField] private ScriptMachine interactionScriptMachine = null;
 
         [Header("Allowed states")]
         [SerializeField] private EAllowedGenericInteractableState desktopAllowedStates = EAllowedGenericInteractableState.Selected | EAllowedGenericInteractableState.Interacting;
         [SerializeField] private EAllowedGenericInteractableState vrAllowedStates = EAllowedGenericInteractableState.Selected | EAllowedGenericInteractableState.Interacting;
 
+        public Action<GameObject> OnSelectedActionVisualScripting;
 
         public List<AwaitableScriptableAction> OnHoverEnterActions { get => onHoverEnterActions; set => onHoverEnterActions = value; }
         public List<AwaitableScriptableAction> OnHoverExitActions { get => onHoverExitActions; set => onHoverExitActions = value; }
@@ -56,6 +58,8 @@ namespace Reflectis.SDK.InteractionNew
         public List<AwaitableScriptableAction> OnDeselectedActions { get => onDeselectedActions; set => onDeselectedActions = value; }
         public List<AwaitableScriptableAction> OnInteractActions { get => onInteractActions; set => onInteractActions = value; }
         public List<AwaitableScriptableAction> OnInteractFinishActions { get => onInteractFinishActions; set => onInteractFinishActions = value; }
+
+        public ScriptMachine InteractionScriptMachine { get => interactionScriptMachine; set => interactionScriptMachine = value; }
 
         public EAllowedGenericInteractableState DesktopAllowedStates { get => desktopAllowedStates; set => desktopAllowedStates = value; }
         public EAllowedGenericInteractableState VRAllowedStates { get => vrAllowedStates; set => vrAllowedStates = value; }
@@ -77,12 +81,31 @@ namespace Reflectis.SDK.InteractionNew
             }
         }
 
+
         private bool hasHoveredState = false;
         private bool skipSelectState = false;
         private bool hasInteractState = false;
 
+        private List<HoverEnterEventUnit> hoverEnterEventUnits = new List<HoverEnterEventUnit>();
 
-        private void Awake()
+        private List<HoverExitEventUnit> hoverExitEventUnits = new List<HoverExitEventUnit>();
+
+        private List<SelectEnterEventUnit> selectEnterEventUnits = new List<SelectEnterEventUnit>();
+
+        private List<SelectExitEventUnit> selectExitEventUnits = new List<SelectExitEventUnit>();
+
+        private List<InteractEventUnit> interactEventUnits = new List<InteractEventUnit>();
+
+        private void OnDestroy()
+        {
+            if (!IsIdleState && CurrentInteractionState != EGenericInteractableState.SelectExiting)
+            {
+                onDeselectingActions.ForEach(x => x.Action());
+                onDeselectedActions.ForEach(x => x.Action());
+            }
+        }
+
+        public override Task Setup()
         {
             switch (SM.GetSystem<IPlatformSystem>().RuntimePlatform)
             {
@@ -97,33 +120,65 @@ namespace Reflectis.SDK.InteractionNew
                     hasHoveredState = VRAllowedStates.HasFlag(EAllowedGenericInteractableState.Hovered);
                     break;
             }
-        }
 
-        private void OnDestroy()
-        {
-            if (!IsIdleState && CurrentInteractionState != EGenericInteractableState.SelectExiting)
+            if (interactionScriptMachine != null)
             {
-                onDeselectingActions.ForEach(x => x.Action());
-                onDeselectedActions.ForEach(x => x.Action());
+                foreach (var unit in interactionScriptMachine.graph.units)
+                {
+                    if (unit is HoverEnterEventUnit hoverEnterEventUnit)
+                    {
+                        hoverEnterEventUnits.Add(hoverEnterEventUnit);
+                    }
+                    if (unit is HoverExitEventUnit hoverExitEventUnit)
+                    {
+                        hoverExitEventUnits.Add(hoverExitEventUnit);
+                    }
+                    if (unit is SelectEnterEventUnit selectEnterEventUnit)
+                    {
+                        selectEnterEventUnits.Add(selectEnterEventUnit);
+                    }
+                    if (unit is SelectExitEventUnit selectExitEventUnit)
+                    {
+                        selectExitEventUnits.Add(selectExitEventUnit);
+                    }
+                    if (unit is InteractEventUnit interactEventUnit)
+                    {
+                        interactEventUnits.Add(interactEventUnit);
+                    }
+                }
             }
+            return Task.CompletedTask;
         }
 
-        public override abstract Task Setup();
-
-        public override void OnHoverStateEntered()
+        public override async void OnHoverStateEntered()
         {
             //if (!CanInteract || !hasHoveredState)
             if (CurrentBlockedState != 0 || !hasHoveredState)
                 return;
+
+            IEnumerable<Task> hoverEnterUnitsTask = hoverEnterEventUnits.Select(async unit =>
+            {
+                await unit.AwaitableTrigger(interactionScriptMachine.GetReference().AsReference(), this);
+            });
+
+            await Task.WhenAll(hoverEnterUnitsTask);
 
             onHoverEnterActions.ForEach(a => a.Action(InteractableRef));
+
         }
 
-        public override void OnHoverStateExited()
+        public override async void OnHoverStateExited()
         {
             //if (!CanInteract || !hasHoveredState)
             if (CurrentBlockedState != 0 || !hasHoveredState)
                 return;
+
+            IEnumerable<Task> hoverExitUnitsTask = hoverExitEventUnits.Select(async unit =>
+            {
+                await unit.AwaitableTrigger(interactionScriptMachine.GetReference().AsReference(), this);
+            });
+
+            await Task.WhenAll(hoverExitUnitsTask);
 
             onHoverExitActions.ForEach(a => a.Action(InteractableRef));
         }
@@ -141,6 +196,13 @@ namespace Reflectis.SDK.InteractionNew
             {
                 await action.Action(InteractableRef);
             }
+
+            IEnumerable<Task> selectEnterUnitsTasks = selectEnterEventUnits.Select(async unit =>
+            {
+                await unit.AwaitableTrigger(interactionScriptMachine.GetReference().AsReference(), this);
+            });
+
+            await Task.WhenAll(selectEnterUnitsTasks);
 
             CurrentInteractionState = EGenericInteractableState.Selected;
             foreach (var action in onSelectedActions)
@@ -168,6 +230,13 @@ namespace Reflectis.SDK.InteractionNew
                 await action.Action(InteractableRef);
             }
 
+            IEnumerable<Task> selectExitUnitsTasks = selectExitEventUnits.Select(async unit =>
+            {
+                await unit.AwaitableTrigger(interactionScriptMachine.GetReference().AsReference(), this);
+            });
+
+            await Task.WhenAll(selectExitUnitsTasks);
+
             CurrentInteractionState = EGenericInteractableState.Idle;
             foreach (var action in OnDeselectedActions)
             {
@@ -189,6 +258,13 @@ namespace Reflectis.SDK.InteractionNew
             {
                 await action.Action(InteractableRef);
             }
+
+            IEnumerable<Task> interactUnitsTasks = interactEventUnits.Select(async unit =>
+            {
+                await unit.AwaitableTrigger(interactionScriptMachine.GetReference().AsReference(), this);
+            });
+
+            await Task.WhenAll(interactUnitsTasks);
 
             CurrentInteractionState = EGenericInteractableState.Selected;
             foreach (var action in onInteractFinishActions)