+++ InteractionNew/Runtime/Scripts/VisualScripting/SelectEnterEventUnit.cs
@@ -1,0 +1,11 @@
+using Unity.VisualScripting;
+
+namespace Reflectis.SDK.InteractionNew
+{
+    [UnitTitle("OnInteractableSelectEnter")]//Custom EventUnit to receive the event. Adding On to the unit title as an event naming convention.
+    [UnitCategory("Events\\ReflectisEvent\\GenericInteractable")]
+    public class SelectEnterEventUnit : GenericInteractableEventUnit
+    {
+
+    }
+}