+++ Utilities/Runtime/Scripts/ImageDownloader.cs
@@ -10,19 +10,13 @@ namespace Reflectis.SDK.Utilities
     {
         public static Dictionary<string, Texture2D> userIconCached = new Dictionary<string, Texture2D>();
 
-        private static Worker worker;
-
         public static void DownloadImage(string mediaUrl, Action<Texture2D> onCompletionCallback, Action onFailedCallback = null, string key = null)
         {
             if (string.IsNullOrEmpty(key))
             {
                 key = mediaUrl;
             }
-            if (worker == null)
-            {
-                GameObject go = new GameObject("ImageDownloaderWorker");
-                worker = go.AddComponent<Worker>();
-            }
+            
             if (string.IsNullOrEmpty(key))
             {
                 Debug.LogError("Trying to download an image with an empty url!");
@@ -36,12 +30,12 @@ namespace Reflectis.SDK.Utilities
                 }
                 else
                 {
-                    worker.StartCoroutine(WaitForTextureDownload(mediaUrl, onCompletionCallback, onFailedCallback, key));
+                    CoroutineRunner.Instance.StartCoroutine(WaitForTextureDownload(mediaUrl, onCompletionCallback, onFailedCallback, key));
                 }
             }
             else
             {
-                worker.StartCoroutine(DownloadImageCoroutine(mediaUrl, onCompletionCallback, onFailedCallback, key));
+                CoroutineRunner.Instance.StartCoroutine(DownloadImageCoroutine(mediaUrl, onCompletionCallback, onFailedCallback, key));
             }
         }